Addressing review.
